Step2 코드리뷰반영 3 StringCalculatorValidator 추가- validation 책임 분리
